sw64: cpufreq: move cpufreq notifier to its user
Move cpufreq notifier to its user to make the code clean without any
functional changes.
