+++ arch/sw_64/include/asm/cpufreq.h
@@ -83,7 +83,5 @@ int sw64_set_rate(unsigned int index);
 
 struct clk *sw64_clk_get(struct device *dev, const char *id);
 
-void sw64_update_clockevents(unsigned long cpu, u32 freq);
-
 unsigned int __sw64_cpufreq_get(struct cpufreq_policy *policy);
 #endif /* _ASM_SW64_CPUFREQ_H */
